% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggDensityJitter.R
\name{ggDensityJitter}
\alias{ggDensityJitter}
\title{wrapper for density plot with jittered x points using ggRidges package}
\usage{
ggDensityJitter(.data, title = "", xlab = "", ylab = "",
  scaled = TRUE, .jitter.width = 0.05, .alpha = 0.3, .scale = 0.9,
  .verbose = TRUE, ...)
}
\arguments{
\item{.data}{a data frame or a matrix}

\item{title}{title text passed to labs(title = title, ...)}

\item{xlab}{text passed to labs(xlab = xlab, ...)}

\item{ylab}{text passed to labs(ylab = ylab, ...)}

\item{scaled}{logical. If .data are scaled for each colmn (default = TRUE).}

\item{.jitter.width}{numeric. Width for horizonal jittering. By default set to 0.05.}

\item{.alpha}{numeric (0, 1). Alpha transparency scales for density plots.}

\item{.scale}{numeric. Position scales for Y-axis.}

\item{.verbose}{ligical. If TRUE, which colmn were converted (from factor to numeric) is printed.}

\item{...}{other options passed to theme(., ...)}
}
\value{
a ggplots object
}
\description{
template wrapper ggridges::geom_density_ridges(., jittered_points = TRUE, position = position_points_jitter())
}
\examples{
\dontrun{
ggDensityJitter(iris)

iris \%>\%
  select(-Species) \%>\%
  ggDensityJitter("feature distribution", .scale = 0.7, .jitter.width = 0.1, .verbose = FALSE)
}
}
