% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggRaincoud.R
\name{ggRaincloud}
\alias{ggRaincloud}
\title{rain could plot wrapper for ggRidges package}
\usage{
ggRaincloud(.data, title = "", xlab = "", ylab = "", scaled = TRUE,
  .alpha = 0.3, .scale = 0.9, .verbose = TRUE, ...)
}
\arguments{
\item{.data}{a data frame or a matrix}

\item{title}{title text passed to labs(title = title, ...)}

\item{xlab}{text passed to labs(xlab = xlab, ...)}

\item{ylab}{text passed to labs(ylab = ylab, ...)}

\item{scaled}{logical. If .data are scaled for each colmn (default = TRUE).}

\item{.alpha}{numeric (0, 1). Alpha transparency scales for density plots.}

\item{.scale}{numeric. Position scales for Y-axis.}

\item{.verbose}{ligical. If TRUE, which colmn were converted (from factor to numeric) is printed.}

\item{...}{other options passed to theme(., ...)}
}
\value{
a ggplots object
}
\description{
template wrapper ggridges::geom_density_ridges(., jittered_points = TRUE, position = "raincloud")
}
\examples{
\dontrun{
ggRaincloud(iris)

iris \%>\%
  select(-Species) \%>\%
  ggRaincloud("feature distribution",.scale = 0.7 .verbose = FALSE)
}
}
