% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggumap-package.R
\docType{package}
\name{ggumap-package}
\alias{ggumap}
\alias{ggumap-package}
\title{ggumap: ggplot interfases for Uniform Manifold Approximation and Projection (UMAP)}
\description{
ggplot interfase for Uniform Manifold Approximation and Projection (UMAP).
    Provides 2D plot for UMAP, and adding clustered-labels based on hierarchical Density-based spatial clustering of applications with noise (HDBSCAN).
}
\keyword{internal}
