% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRdata.R
\docType{data}
\name{HRdata}
\alias{HRdata}
\title{Human Resource dataset}
\format{A data frame with 14999 rows of 10 variables:
\describe{
\item{satisfaction_level}{ Level of satisfaction (0-1) }
\item{last_evaluation}{ Time since last performance evaluation (in Years) }
\item{number_project}{ Number of projects completed while at work }
\item{average_montly_hours}{ Average monthly hours at workplace }
\item{time_spend_company}{ Number of years spent in the company }
\item{Work_accident}{ Whether the employee had a workplace accident }
\item{left}{ Whether the employee left the workplace or not (1 or 0) Factor }
\item{promotion_last_5years}{ Whether the employee was promoted in the last five years }
\item{sales}{ Department in which they work for }
\item{salary}{ Relative level of salary (high) }
\item{}{}
}}
\source{
\url{https://github.com/ryankarlos/Human-Resource-Analytics-Kaggle-Dataset/tree/master/Original_Kaggle_Dataset}
}
\usage{
HRdata
}
\description{
A dataset from Kaggle competition Human Resources Analytics: Why are our best and most experienced employees leaving prematurely? according to \code{help(breakDown::HR_data)}.
}
\keyword{datasets}
